Move output to bin
